[308]: Splitting mapper into several sub-interfaces for organisational purposes
