+++ src/test/java/uk/ac/cam/cl/dtg/segue/api/managers/UserAccountManagerTest.java
@@ -45,8 +45,8 @@ import uk.ac.cam.cl.dtg.segue.dao.schools.UnableToIndexSchoolsException;
 import uk.ac.cam.cl.dtg.segue.dao.users.PgAnonymousUsers;
 import uk.ac.cam.cl.dtg.segue.dao.users.PgUsers;
 import uk.ac.cam.cl.dtg.segue.search.SegueSearchException;
-import uk.ac.cam.cl.dtg.util.MapStructContentMapper;
 import uk.ac.cam.cl.dtg.util.PropertiesLoader;
+import uk.ac.cam.cl.dtg.util.mappers.MapStructMainMapper;
 
 class UserAccountManagerTest {
 
@@ -62,7 +62,7 @@ class UserAccountManagerTest {
     PropertiesLoader propertiesLoader = createNiceMock(PropertiesLoader.class);
     Map<AuthenticationProvider, IAuthenticator> providersToRegister = Map.of();
     MapperFacade dtoMapper = createNiceMock(MapperFacade.class);
-    MapStructContentMapper newMapper = MapStructContentMapper.INSTANCE;
+    MapStructMainMapper newMapper = MapStructMainMapper.INSTANCE;
     emailManager = createMock(EmailManager.class);
     PgAnonymousUsers pgAnonymousUsers = createNiceMock(PgAnonymousUsers.class);
     ILogManager logManager = createNiceMock(ILogManager.class);
